+++ lib/easy_table.dart
@@ -17,3 +17,4 @@ export 'src/row_hover_listener.dart';
 export 'src/row_callbacks.dart';
 export 'src/cell_style.dart';
 export 'src/cell_icon.dart';
+export 'src/last_visible_row_listener.dart';