+++ lib/src/last_visible_row_listener.dart
@@ -1,0 +1,2 @@
+/// Signature for listening to the last visible row index.
+typedef OnLastVisibleRowListener = void Function(int lastVisibleRowIndex);